requirements.txt
matplotlib
librosa==0.9.1
tqdm
pandas
numba==0.53.1
numpy==1.23.5
scipy
PyYAML
tensorboardX
pyloudnorm
espnet
setuptools>=41.0.0
g2p_en
resemblyzer
webrtcvad
tensorboard==2.6.0
scikit-learn==0.24.1
scikit-image==0.16.2
textgrid
jiwer
pycwt
PyWavelets
praat-parselmouth==0.3.3
jieba
einops
chardet
protobuf==3.20.*
pyopenjtalk
utaupy
music21
mido
ffmpeg-normalize
soundcard
logger
torchcrepe
